requirements.txt
aiogram
aiohttp
motor
pymongo
dnspython
python-dotenv
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
beautifulsoup4
lxml
msal
msgraph-sdk
